归一化版本合并
版本合并

+++ elfin_ethercat_driver/include/elfin_ethercat_driver/elfin_ethercat_manager.h
@@ -44,7 +44,7 @@ namespace elfin_ethercat_driver {
 typedef struct {
     std::string name;
     int32_t value;
-    uint8_t channel;
+    uint16_t channel;
 }ElfinPDOunit;
 
 /**
